backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.29.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pandas>=2.2.0
